single script for tensorflow models
