+++ configs/audioset-yamnet-1.json
@@ -1,0 +1,4 @@
+{
+    "model_name": "audioset-yamnet-1",
+    "sample_rate": 16000
+}
