+++ pipeline_experiment_scripts/yamnet-search_evaluate_experiments.sh
@@ -7,7 +7,7 @@ export PATH="$SCRIPT_DIR:$PATH"
 
 DATA_DIR="/home/roguz/freesound/freesound-perceptual_similarity/data"
 DATASET_NAME="FSD50K.eval_audio"
-MODEL_NAME="audioset-yamnet_v1"
+MODEL_NAME="audioset-yamnet-1"
 
 EMBED_DIR="$DATA_DIR/embeddings/$DATASET_NAME"
