+++ tensorflow_predict-embedding_extractor.py
@@ -1,5 +1,5 @@
 """Takes a FSD50K csv file specifying audio file names and computes embeddings 
-using audioset-yamnet_v1. All frame embeddings are exported without aggregation."""
+using FSD-Sinet. All frame embeddings are exported without aggregation."""
 
 import os
 import time
@@ -9,15 +9,11 @@ from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 import numpy as np
 import pandas as pd
 
-from essentia.standard import EasyLoader, TensorflowPredictVGGish
+from essentia.standard import EasyLoader, TensorflowPredictFSDSINet, TensorflowPredictVGGish
 
 from directories import AUDIO_DIR, GT_PATH, EMBEDDINGS_DIR, MODELS_DIR
 
-TRIM_DUR = 30
-SAMPLE_RATE = 16000
-ANALYZER_NAME = 'audioset-yamnet-1'
-MODEL_PATH = f"{MODELS_DIR}/{ANALYZER_NAME}.pb"
-EMBEDDINGS_DIR = f"{EMBEDDINGS_DIR}/{ANALYZER_NAME}"
+TRIM_DUR = 30 # seconds
 
 # TODO: only discard non-floatable frames?
 def create_embeddings(model, audio):
@@ -32,15 +28,15 @@ def create_embeddings(model, audio):
 
 # TODO: effect of zero padding short clips?
 # TODO: energy based frame filtering (at audio input)
-def process_audio(model_embeddings, audio_path, output_dir):
+def process_audio(model_embeddings, audio_path, output_dir, sample_rate):
     """ Reads the audio of given path, creates the embeddings and exports."""
     # Load the audio file
     loader = EasyLoader()
-    loader.configure(filename=audio_path, sampleRate=SAMPLE_RATE, endTime=TRIM_DUR, replayGain=0)
+    loader.configure(filename=audio_path, sampleRate=sample_rate, endTime=TRIM_DUR, replayGain=0)
     audio = loader()
     # Zero pad short clips
-    if audio.shape[0] < SAMPLE_RATE:
-        audio = np.concatenate((audio, np.zeros((SAMPLE_RATE-audio.shape[0]))))
+    if audio.shape[0] < sample_rate:
+        audio = np.concatenate((audio, np.zeros((sample_rate-audio.shape[0]))))
     # Process
     embeddings = create_embeddings(model_embeddings, audio)
     # Save results
@@ -54,13 +50,28 @@ if __name__=="__main__":
     parser=ArgumentParser(description=__doc__, 
                                    formatter_class=ArgumentDefaultsHelpFormatter)
     parser.add_argument('-p', '--path', type=str, default=GT_PATH, 
-                        help='Path to csv file containing fnames.')
+                        help='Path to csv file containing audio fnames.')
+    parser.add_argument('-c', '--config', type=str, required=True,
+                        help="Path to config file of the model.")
+    parser.add_argument('-o', '--output_dir', type=str, default=EMBEDDINGS_DIR,
+                        help="Path to output directory.")
     args=parser.parse_args()
 
+    # Read the config file
+    with open(args.config, "r") as json_file:
+        config = json.load(json_file)
+
     # Configure the embedding model
-    model_embeddings = TensorflowPredictVGGish(graphFilename=MODEL_PATH, 
-                                               input="melspectrogram", 
-                                               output="embeddings")
+    model_path = os.path.join(MODELS_DIR, f"{config['model_name']}.pb")
+    if "audioset-yamnet" in config['model_name']:
+        model_embeddings = TensorflowPredictVGGish(graphFilename=model_path, 
+                                                input="melspectrogram", 
+                                                output="embeddings")
+    elif "fsd-sinet" in config['model_name']:
+        model_embeddings = TensorflowPredictFSDSINet(graphFilename=model_path,
+                                                    output="model/global_max_pooling1d/Max")
+    else:
+        raise ValueError(f"Unknown model name: {config['model_name']}")
 
     # Read the file names
     fnames = pd.read_csv(args.path)["fname"].to_list()
@@ -68,15 +79,16 @@ if __name__=="__main__":
     print(f"There are {len(audio_paths)} audio files to process.")
 
     # Create the output directory
-    os.makedirs(EMBEDDINGS_DIR, exist_ok=True)
-    print(f"Exporting the embeddings to: {EMBEDDINGS_DIR}")
+    output_dir = os.path.join(args.output_dir, config['model_name'])
+    os.makedirs(output_dir, exist_ok=True)
+    print(f"Exporting the embeddings to: {output_dir}")
 
     # Process each audio
     start_time = time.time()
     for i,audio_path in enumerate(audio_paths):
         if i%1000==0:
             print(f"[{i:>{len(str(len(audio_paths)))}}/{len(audio_paths)}]")
-        process_audio(model_embeddings, audio_path, EMBEDDINGS_DIR)
+        process_audio(model_embeddings, audio_path, output_dir, config['sample_rate'])
     total_time = time.time()-start_time
     print(f"\nTotal time: {time.strftime('%M:%S', time.gmtime(total_time))}")
     print(f"Average time/file: {total_time/len(audio_paths):.2f} sec.")
